test(tools): split observability payload coverage
